# api/openapispec/swagger.yaml
definitions:
...
  entity.ResourceGroup:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      apiVersion:
        type: string
      cluster:
        type: string
      kind:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
    type: object
...
  insight.ScoreData:
    properties:
      issuesTotal:
        description: |-
          IssuesTotal is the total count of all issues found during the audit.
          This count can be used to understand the overall number of problems
          that need to be addressed.
        type: integer
      resourceTotal:
        description: ResourceTotal is the count of unique resources audited during
          the scan.
        type: integer
      score:
        description: |-
          Score represents the calculated score of the audited manifest based on
          the number and severity of issues. It provides a quantitative measure
          of the security posture of the resources in the manifest.
        type: number
      severityStatistic:
        additionalProperties:
          type: integer
        description: |-
          SeverityStatistic is a mapping of severity levels to their respective
          number of occurrences. It allows for a quick overview of the distribution
          of issues across different severity categories.
        type: object
    type: object
  insight.Statistics:
    properties:
      clusterCount:
        type: integer
      resourceCount:
        type: integer
      resourceGroupRuleCount:
        type: integer
    type: object
...
  scanner.AuditData:
    properties:
      bySeverity:
        additionalProperties:
          type: integer
        type: object
      issueGroups:
        items:
          $ref: '#/definitions/scanner.IssueGroup'
        type: array
      issueTotal:
        type: integer
      resourceTotal:
        type: integer
    type: object
  scanner.Issue:
    properties:
      message:
        description: Message provides a detailed human-readable description of the
          issue.
        type: string
      scanner:
        description: Scanner is the name of the scanner that discovered the issue.
        type: string
      severity:
        description: Severity indicates how critical the issue is, using the IssueSeverityLevel
          constants.
        type: integer
      title:
        description: Title is a brief summary of the issue.
        type: string
    type: object
  scanner.IssueGroup:
    properties:
      issue:
        $ref: '#/definitions/scanner.Issue'
      resourceGroups:
        items:
          $ref: '#/definitions/entity.ResourceGroup'
        type: array
    type: object
...
info:
  contact: {}
  title: Karpor
  version: "1.0"
paths:
...
  /rest-api/v1/insight/audit:
    get:
      description: This endpoint audits based on the specified resource group.
      parameters:
      - description: The specified cluster name, such as 'example-cluster'
        in: query
        name: cluster
        type: string
      - description: The specified apiVersion, such as 'apps/v1'
        in: query
        name: apiVersion
        type: string
      - description: The specified kind, such as 'Deployment'
        in: query
        name: kind
        type: string
      - description: The specified namespace, such as 'default'
        in: query
        name: namespace
        type: string
      - description: The specified resource name, such as 'foo'
        in: query
        name: name
        type: string
      - description: Switch for forced scanning, default is 'false'
        in: query
        name: forceNew
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Audit results
          schema:
            $ref: '#/definitions/scanner.AuditData'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Audit based on resource group.
      tags:
      - insight
...
  /rest-api/v1/insight/score:
    get:
      description: This endpoint calculates a score for the provided manifest based
        on the number and severity of issues detected during the audit.
      parameters:
      - description: The specified cluster name, such as 'example-cluster'
        in: query
        name: cluster
        type: string
      - description: The specified apiVersion, such as 'apps/v1'
        in: query
        name: apiVersion
        type: string
      - description: The specified kind, such as 'Deployment'
        in: query
        name: kind
        type: string
      - description: The specified namespace, such as 'default'
        in: query
        name: namespace
        type: string
      - description: The specified resource name, such as 'foo'
        in: query
        name: name
        type: string
      - description: Switch for forced compute score, default is 'false'
        in: query
        name: forceNew
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Score calculation result
          schema:
            $ref: '#/definitions/insight.ScoreData'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: ScoreHandler calculates a score for the audited manifest.
      tags:
      - insight
  /rest-api/v1/insight/stats:
    get:
      description: This endpoint returns a global statistics info.
      produces:
      - application/json
      responses:
        "200":
          description: Global statistics info
          schema:
            $ref: '#/definitions/insight.Statistics'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get returns a global statistics info.
      tags:
      - insight
...
  /rest-api/v1/search:
    get:
      description: This endpoint returns an array of Kubernetes runtime Object matched
        using the query from context.
      parameters:
      - description: The query to use for search. Required
        in: query
        name: query
        required: true
        type: string
      - description: The search pattern. Can be either sql or dsl. Required
        in: query
        name: pattern
        required: true
        type: string
      - description: The size of the page. Default to 10
        in: query
        name: pageSize
        type: string
      - description: The current page to fetch. Default to 1
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of runtime.Object
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: SearchForResource returns an array of Kubernetes runtime Object matched
        using the query from context.
      tags:
      - search
swagger: "2.0"